Fixed evaluation and defaults.

+++ src/pl_modules/losses.py
@@ -15,7 +15,8 @@ def cce(input, target, maxval=None, weights=None):
     #     weights = weights.reshape(1, weights_len, 1, 1)
     loss = nn.CrossEntropyLoss()
     # target = torch.argmax(target, 1)
-    output = loss(input.float(), target.float().squeeze(1))
+    # output = loss(input.float(), target.float().squeeze(1))
+    output = loss(input.float(), target.long().squeeze(1))
     return output
